requirements.txt
Flask
Flask-restful
Flask-SQLAlchemy
mysqlclient
python-dotenv